refactor(frontend): tighten orchestrator seams
