+++ packages/frontend/src/state/operations/agent-orchestrator/hooks/use-repo-session-hydration-effects.test.tsx
@@ -1,0 +1,43 @@
+import { describe, expect, test } from "bun:test";
+import { createHookHarness } from "@/test-utils/react-hook-harness";
+import { AgentSessionPresenceStore } from "../lifecycle/session-presence-store";
+import { taskWithSession } from "./orchestrator-hook-test-fixtures";
+import { useRepoSessionHydrationEffects } from "./use-repo-session-hydration-effects";
+
+(globalThis as unknown as { IS_REACT_ACT_ENVIRONMENT: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
+
+describe("useRepoSessionHydrationEffects", () => {
+  test("bootstraps and reconciles pending task sessions", async () => {
+    const loadCalls: Array<{ taskId: string; mode: string | undefined }> = [];
+    const Harness = ({ repoPath }: { repoPath: string | null }) =>
+      useRepoSessionHydrationEffects({
+        workspaceRepoPath: repoPath,
+        tasks: repoPath ? [taskWithSession] : [],
+        currentWorkspaceRepoPathRef: { current: repoPath },
+        agentSessionPresenceStore: new AgentSessionPresenceStore(),
+        sessionHydration: {
+          bootstrapTaskSessions: async (taskId) => {
+            loadCalls.push({ taskId, mode: undefined });
+          },
+          reconcileLiveTaskSessions: async ({ taskId }) => {
+            loadCalls.push({ taskId, mode: "reconcile_live" });
+          },
+        },
+      });
+    const harness = createHookHarness<{ repoPath: string | null }, ReturnType<typeof Harness>>(
+      Harness,
+      { repoPath: "/tmp/repo" },
+    );
+    await harness.mount();
+    await harness.waitFor(() => loadCalls.length >= 2);
+
+    expect(loadCalls).toContainEqual({ taskId: "task-1", mode: undefined });
+    expect(loadCalls).toContainEqual({ taskId: "task-1", mode: "reconcile_live" });
+
+    await harness.update({ repoPath: null });
+    const countAfterRepoReset = loadCalls.length;
+    await Promise.resolve();
+    expect(loadCalls).toHaveLength(countAfterRepoReset);
+    await harness.unmount();
+  });
+});
