+++ packages/frontend/src/state/operations/agent-orchestrator/lifecycle/load-sessions.test.ts
@@ -30,14 +30,8 @@ import { AgentSessionPresenceStore } from "./session-presence-store";
 
 type AgentSessionState = BaseAgentSessionState & { runId?: string | null };
 
-const createLoadAgentSessions = (
-  args: Omit<Parameters<typeof createLoadAgentSessionsBase>[0], "queryClient"> & {
-    queryClient?: Parameters<typeof createLoadAgentSessionsBase>[0]["queryClient"];
-  },
-) => {
-  const { queryClient = appQueryClient, ...rest } = args;
-  return createLoadAgentSessionsBase({ ...rest, queryClient });
-};
+const createLoadAgentSessions = (args: Parameters<typeof createLoadAgentSessionsBase>[0]) =>
+  createLoadAgentSessionsBase(args);
 
 type LegacyRunSummary = { runId: string; worktreePath: string };
 
@@ -262,6 +256,7 @@ describe("agent-orchestrator-load-sessions", () => {
   test("no-ops when active repo is missing", async () => {
     let setCalled = false;
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: null,
       adapter: createAdapter(),
       repoEpochRef: { current: 0 },
@@ -283,6 +278,7 @@ describe("agent-orchestrator-load-sessions", () => {
   test("no-ops for blank task ids", async () => {
     let setCalled = false;
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -333,6 +329,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -467,6 +464,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -642,6 +640,7 @@ describe("agent-orchestrator-load-sessions", () => {
     ]);
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -784,6 +783,7 @@ describe("agent-orchestrator-load-sessions", () => {
     ]);
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -933,6 +933,7 @@ describe("agent-orchestrator-load-sessions", () => {
       externalSessionId: string;
     } | null = null;
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -1079,6 +1080,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -1211,6 +1213,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -1325,6 +1328,7 @@ describe("agent-orchestrator-load-sessions", () => {
     const observedSnapshotDirectories: string[][] = [];
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -1438,6 +1442,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -1584,6 +1589,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -1862,6 +1868,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -2009,6 +2016,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -2140,6 +2148,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -2299,6 +2308,7 @@ describe("agent-orchestrator-load-sessions", () => {
     );
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -2388,6 +2398,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
     let historyLoads = 0;
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -2485,6 +2496,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -2633,6 +2645,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -2729,6 +2742,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -2866,6 +2880,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -2951,6 +2966,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -3092,6 +3108,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -3167,6 +3184,7 @@ describe("agent-orchestrator-load-sessions", () => {
       },
     };
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -3286,6 +3304,7 @@ describe("agent-orchestrator-load-sessions", () => {
 
     let historyLoads = 0;
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -3456,6 +3475,7 @@ describe("agent-orchestrator-load-sessions", () => {
     });
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -3568,6 +3588,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -3644,6 +3665,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -3738,6 +3760,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -3860,6 +3883,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -3961,6 +3985,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -4077,6 +4102,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -4179,6 +4205,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -4328,6 +4355,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -4440,6 +4468,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -4548,6 +4577,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -4660,6 +4690,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -4784,6 +4815,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -4870,6 +4902,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -5004,6 +5037,7 @@ describe("agent-orchestrator-load-sessions", () => {
       createDeferred<Awaited<ReturnType<ReturnType<typeof createAdapter>["loadSessionHistory"]>>>();
     let historyCalls = 0;
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -5123,6 +5157,7 @@ describe("agent-orchestrator-load-sessions", () => {
       createDeferred<Awaited<ReturnType<ReturnType<typeof createAdapter>["loadSessionHistory"]>>>();
     let historyCalls = 0;
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -5288,6 +5323,7 @@ describe("agent-orchestrator-load-sessions", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
+      queryClient: appQueryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
