+++ packages/frontend/src/state/operations/agent-orchestrator/support/orchestrator-dependencies.ts
@@ -8,8 +8,6 @@ import { appQueryClient } from "@/lib/query-client";
 import { host } from "../../shared/host";
 
 export type AgentOrchestratorHostPort = {
-  buildStart: typeof host.buildStart;
-  runtimeEnsure: typeof host.runtimeEnsure;
   agentSessionUpsert: (
     repoPath: string,
     taskId: string,
@@ -19,21 +17,29 @@ export type AgentOrchestratorHostPort = {
   taskWorktreeGet: (repoPath: string, taskId: string) => Promise<TaskWorktreeSummary | null>;
 };
 
+export type AgentOrchestratorRuntimeHostPort = {
+  buildStart: typeof host.buildStart;
+  runtimeEnsure: typeof host.runtimeEnsure;
+};
+
 export type AgentOrchestratorDependencies = {
   queryClient: QueryClient;
   hostPort: AgentOrchestratorHostPort;
+  runtimeHostPort: AgentOrchestratorRuntimeHostPort;
 };
 
 export const createDefaultAgentOrchestratorDependencies = (): AgentOrchestratorDependencies => ({
   queryClient: appQueryClient,
   hostPort: {
-    buildStart: (...args) => host.buildStart(...args),
-    runtimeEnsure: (...args) => host.runtimeEnsure(...args),
     agentSessionUpsert: (repoPath, taskId, record) =>
       host.agentSessionUpsert(repoPath, taskId, record),
     agentSessionStop: async (target) => {
       await host.agentSessionStop(target);
     },
     taskWorktreeGet: (repoPath, taskId) => host.taskWorktreeGet(repoPath, taskId),
+  },
+  runtimeHostPort: {
+    buildStart: (...args) => host.buildStart(...args),
+    runtimeEnsure: (...args) => host.runtimeEnsure(...args),
   },
 });
